Phil: Please pull down the changes and note what I have done. You should see the input in the Settings > Data Inputs > Chatchpoint Modular Input in the Splunk UI.
Committing modular input that I created. I have made some edits to the structure. With the consumer_key and consumer_secret stored in the setup page then there is only a need for the test_id to be in the modular input. I have made the changes accordingly.

This commit is a commit only for the purpose of the adjustment in the structure of the modular input input fields.

Issues: having issues indexing data into splunk. This will take some debugging but shouldn't be too difficult to work out this bug.